refac: Update harbor package and health check
